+++ packages/orq-rc/RELEASES.md
@@ -5999,3 +5999,13 @@ Based on:
 - [typescript v4.2.0-rc.46] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.46 - packages/orq-rc
+
+## 2026-01-20 11:21:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.690.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.47] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.47 - packages/orq-rc